Klocwork fixes for complaints about casting
DbConfig:  Passing a size_t to an int.  Changed function to use unsigned int as currentLineNumber should never be negative.

Antenna:  Changed freg to float as function call expected a float

http://simapache.enews.nrl.navy.mil/reviews/r/27646/

+++ SDK/simVis/Antenna.cpp
@@ -170,7 +170,7 @@ bool AntennaNode::setPrefs(const simData::BeamPrefs& prefs)
     // load the new pattern file
     delete(antennaPattern_);
     // Frequency must be > 0, if <= 0 use default value
-    const double freq = prefs.frequency() > 0 ? prefs.frequency() : simCore::DEFAULT_FREQUENCY;
+    const float freq = static_cast<float>(prefs.frequency() > 0 ? prefs.frequency() : simCore::DEFAULT_FREQUENCY);
     antennaPattern_ = simCore::loadPatternFile(patternFile_, freq);
     loadedOK_ = (antennaPattern_ != nullptr);
   }
